preliminary user input handling logic
